# python-flask-server-generated/swagger_server/models/completed_task.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class CompletedTask(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, completed: bool=False):  # noqa: E501
        """CompletedTask - a model defined in Swagger

        :param completed: The completed of this CompletedTask.  # noqa: E501
        :type completed: bool
        """
        self.swagger_types = {
            'completed': bool
        }

        self.attribute_map = {
            'completed': 'completed'
        }

        self._completed = completed

    @classmethod
    def from_dict(cls, dikt) -> 'CompletedTask':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CompletedTask of this CompletedTask.  # noqa: E501
        :rtype: CompletedTask
        """
        return util.deserialize_model(dikt, cls)

    @property
    def completed(self) -> bool:
        """Gets the completed of this CompletedTask.


        :return: The completed of this CompletedTask.
        :rtype: bool
        """
        return self._completed

    @completed.setter
    def completed(self, completed: bool):
        """Sets the completed of this CompletedTask.


        :param completed: The completed of this CompletedTask.
        :type completed: bool
        """
        if completed is None:
            raise ValueError("Invalid value for `completed`, must not be `None`")  # noqa: E501

        self._completed = completed